GH-20: Implement factory-like architecture
