+++ package.json
@@ -31,7 +31,7 @@
     "LICENSE"
   ],
   "scripts": {
-    "build": "rm -rf dist && npm run build:cjs && cp src/*.less dist/",
+    "build": "rm -rf dist && npm run build:cjs && cp src/*.less src/*.css dist/",
     "build:cjs": "rollup -c --configPlugin @rollup/plugin-typescript",
     "test": "jest --config jestconfig.json"
   },
@@ -48,7 +48,8 @@
     "@types/jest": "^29.4.0",
     "@types/node": "^18.14.1",
     "@types/react": "^18.0.28",
-    "antd": "^4.24.8",
+    "antd": "npm:antd@^5.3.2",
+    "antd4": "npm:antd@^4.24.8",
     "identity-obj-proxy": "^3.0.0",
     "jest": "^29.4.3",
     "jest-environment-jsdom": "^29.4.3",
